doc-cluster/requirements.txt
hdbscan
openai
pandas
numpy
python-dotenv
tiktoken
notebook
plotly
umap-learn
